+++ src/main/java/com/example/projectwaifu/repositories/InventoryRepository.java
@@ -10,7 +10,7 @@ import java.util.List;
 @Repository
 public interface InventoryRepository extends JpaRepository<UserItem, Long> {
     @Query("SELECT u.productId FROM UserItem u WHERE u.userId = :userid AND u.productId like CONCAT(:productName, '%')")
-    List<String> getPurchased(Long userid, String productName);
+    List<String> getPurchased(Integer userid, String productName);
 
     boolean existsByProductId(String product_id);
 }
